backend/pyproject.toml
[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120
atomic = true
src_paths = ["api"]
extend_skip = ["backend", "api/migrations", ".idea"]

[tool.coverage.run]
source = ["api/"]

[tool.coverage.report]
skip_covered = true
exclude_lines = ["def __str__", "raise NotImplementedError"]

[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Martin Boos <mboos@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.5"
django-cors-headers = "^3.13.0"
djangorestframework = "^3.14.0"
requests = "^2.28.1"
loguru = "^0.6.0"
psycopg2-binary = "^2.9.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {extras = ["d"], version = "^22.12.0"}
isort = "^5.11.4"
coverage = {extras = ["toml"], version = "^7.0.4"}
pre-commit = "^2.21.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
